+++ build-docs.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+set -ev
+
+sphinx-build -n -b html -b linkcheck -d docs/_build/doctrees docs docs/_build/html