+++ docs/conf.py
@@ -23,6 +23,8 @@ sys.path.insert(0, os.path.abspath('..'))
 
 # -- General configuration ------------------------------------------------
 
+suppress_warnings = ['image.nonlocal_uri']
+
 # If your documentation needs a minimal Sphinx version, state it here.
 #needs_sphinx = '1.0'