+++ test.sh
@@ -3,12 +3,12 @@
 
 set -ev
 
-flake8 bidict tests/*.py || { EXIT=1 && echo -e "\0007flake8 failed"; }
-pydocstyle bidict || { EXIT=1 && echo -e "\0007pydocstyle failed"; }
-test -z "$BIDICT_SPHINXBUILD_DISABLE" && { sphinx-build -n -b html -b linkcheck -d docs/_build/doctrees docs docs/_build/html || { EXIT=1 && echo -e "\0007docsbuild failed"; } ;}
+flake8 bidict tests/*.py || { EXIT=1 && echo -e "\033[0;31m * flake8 failed \033[0m \0007"; }
+pydocstyle bidict || { EXIT=1 && echo -e "\033[0;31m * pydocstyle failed \033[0m \0007"; }
+test -z "$BIDICT_BUILD_DOCS_ENABLE" || { ./build-docs.sh || { EXIT=1 && echo -e "\033[0;31m * build-docs.sh failed \033[0m \0007"; } ; }
 
-test -z "$BIDICT_COVERAGE_DISABLE" && COV="--cov=bidict"
-test -n "$BENCHMARK_DIR" && BENCHMARK_STORAGE="--benchmark-storage=$BENCHMARK_DIR"
-py.test $COV $BENCHMARK_STORAGE || { EXIT=1 && echo -e "\0007pytest failed"; }
+test -z "$BIDICT_COVERAGE_ENABLE" || COV="--cov=bidict"
+test -z "$BENCHMARK_DIR" || BENCHMARK_STORAGE="--benchmark-storage=$BENCHMARK_DIR"
+py.test $COV $BENCHMARK_STORAGE $BENCHMARK_SKIP || { EXIT=1 && echo -e "\033[0;31m * pytest failed \033[0m \0007"; }
 
 exit $EXIT