+++ CHANGELOG.rst
@@ -6,9 +6,96 @@ Changelog
 Release Notifications
 ---------------------
 
+.. image:: https://img.shields.io/badge/VersionEye-follow-brightgreen.svg
+    :target: https://www.versioneye.com/python/bidict
+    :alt: Follow on VersionEye
+
 `Follow bidict on VersionEye <https://www.versioneye.com/python/bidict>`_
 to automatically be notified via email
 when a new version of bidict is released.
+
+0.13.0 (not yet released)
+-------------------------
+
+- Support Python 3.6.
+
+- :class:`BidirectionalMapping <bidict.BidirectionalMapping>`
+  has been refactored into an abstract base class,
+  following the way :class:`collections.abc.Mapping` works.
+  The concrete method implementations it used to provide have been moved
+  into a new :class:`BidictBase <bidict.BidictBase>` subclass.
+
+  :class:`BidirectionalMapping <bidict.BidirectionalMapping>`
+  now also implements
+  :attr:`__subclasshook__ <bidict.BidirectionalMapping.__subclasshook__>`,
+  so any class that provides a conforming set of attributes
+  (enumerated in :attr:`_subclsattrs <bidict.BidirectionalMapping._subclsattrs>`)
+  will be considered a
+  :class:`BidirectionalMapping <bidict.BidirectionalMapping>`
+  subclass automatically.
+
+- ``OrderedBidirectionalMapping`` has been renamed to
+  :class:`OrderedBidictBase <bidict.OrderedBidictBase>`,
+  to better reflect its function.
+  (It is not an ABC.)
+
+- A new
+  :class:`FrozenBidictBase <bidict.FrozenBidictBase>` class
+  has been factored out of
+  :class:`frozenbidict <bidict.frozenbidict>` and
+  :class:`frozenorderedbidict <bidict.frozenorderedbidict>`.
+  This implements common behavior such as caching the result of
+  :attr:`__hash__ <bidict.FrozenBidictBase.__hash__>`
+  after the first call.
+
+  It also exposes a new
+  :attr:`_HASH_NITEMS_MAX <bidict.FrozenBidictBase._HASH_NITEMS_MAX>`
+  attribute which can be tuned to bound the time and space complexity when
+  computing the hash.
+
+- :attr:`frozenbidict's __hash__ implementation
+  <bidict.frozenbidict._compute_hash>`
+  and
+  :attr:`frozenorderedbidict's __hash__ implementation
+  <bidict.frozenorderedbidict._compute_hash>`
+  now both create an ephemeral copy of the items to hash
+  (in a :class:`frozenset` and :class:`tuple`, respectively)
+  so as to use the faster ``frozenset_hash`` and ``tuplehash`` routines
+  implemented in C on CPython.
+  It is not possible to use these faster routines with an iterable
+  because there is no such API.
+  Use
+  :attr:`_HASH_NITEMS_MAX <bidict.FrozenBidictBase._HASH_NITEMS_MAX>`
+  to limit the size of the ephemeral copies if they create an issue for you
+  and `report your use case <https://gitter.im/jab/bidict>`_ if possible.
+
+- Make :attr:`frozenorderedbidict's __hash__ implementation
+  <bidict.frozenorderedbidict._compute_hash>` order-sensitive.
+
+  ``hash(frozenorderedbidict([x, y]))`` is now very unlikely to equal
+  ``hash(frozenorderedbidict([y, x]))``.
+  They don't compare equal so their hashes shouldn't be equal either.
+  Avoids hash collisions when inserting such objects into the same set/mapping.
+
+- Add :attr:`_fwd_class <bidict.BidictBase._fwd_class>` and
+  :attr:`_inv_class <bidict.BidictBase._inv_class>` attributes
+  representing the backing :class:`Mapping <collections.abc.Mapping>` types
+  used internally to store the forward and inverse dictionaries, respectively.
+
+  This allows creating custom bidict types with extended functionality
+  simply by overriding these attributes in a subclass.
+
+  See the new :ref:`extending` documentation for examples.
+
+- Pass any parameters passed to :attr:`bidict.popitem <bidict.bidict.popitem>`
+  through to ``_fwd.popitem`` for greater extensibility.
+
+- More concise repr strings for empty bidicts.
+
+  e.g. ``bidict()`` rather than ``bidict({})`` and
+  ``orderedbidict()`` rather than ``orderedbidict([])``.
+
+- Remove unused ``bidict.compat.izip_longest``.
 
 0.12.0 (2016-07-03)
 -------------------
@@ -77,7 +164,7 @@ when a new version of bidict is released.
   - :func:`bidict.compat.iterkeys`
   - :func:`bidict.compat.itervalues`
   - :func:`bidict.compat.izip`
-  - :func:`bidict.compat.izip_longest`
+  - ``bidict.compat.izip_longest``
 
   to complement the existing
   :func:`iteritems() <bidict.compat.iteritems>` and
@@ -87,9 +174,9 @@ when a new version of bidict is released.
 - More efficient implementations of
   :func:`pairs() <bidict.util.pairs>`,
   :func:`inverted() <bidict.util.inverted>`, and
-  :func:`bidict.copy() <bidict.BidirectionalMapping.copy>`.
+  :func:`bidict.copy() <bidict.BidictBase.copy>`.
 
-- Implement :func:`bidict.__copy__() <bidict.BidirectionalMapping.__copy__>`
+- Implement :func:`bidict.__copy__() <bidict.BidictBase.__copy__>`
   for use with the :mod:`copy` module.
 
 - Fix issue preventing a client class from inheriting from
@@ -182,7 +269,7 @@ Breaking API Changes
 ^^^^^^^^^^^^^^^^^^^^
 
 - Remove ``bidict.__invert__``, and with it, support for the ``~b`` syntax.
-  Use :attr:`b.inv <bidict.BidirectionalMapping.inv>` instead.
+  Use :attr:`b.inv <bidict.BidictBase.inv>` instead.
   `#19 <https://github.com/jab/bidict/issues/19>`_
 
 - Remove support for the slice syntax.
@@ -190,7 +277,7 @@ Breaking API Changes
   `#19 <https://github.com/jab/bidict/issues/19>`_
 
 - Remove ``bidict.invert``.
-  Use :attr:`b.inv <bidict.BidirectionalMapping.inv>`
+  Use :attr:`b.inv <bidict.BidictBase.inv>`
   rather than inverting a bidict in place.
   `#20 <https://github.com/jab/bidict/issues/20>`_